+++ src/webapp/src/middleware/fatalError.js
@@ -15,6 +15,19 @@ export class FatalErrorMiddleware {
       this.toasted.global.readonly()
     }
 
+    // Catch generic "Network Error"s that may indicate an Ad Blocker is in use
+    if (
+      !err.response &&
+      err.message === 'Network Error' &&
+      /adwords|ads|analytics/.test(err.config.url)
+    ) {
+      this.toasted.global.error(
+        'The action was blocked, but the browser did not specify why. ' +
+          'If you are using an extension to block ads or tracking tools, it may have been triggered accidentally. ' +
+          'Consider disabling it for this domain, and try again.'
+      )
+    }
+
     throw err
   }
 }